Add files via upload
To run the project, you need to install the dependencies located in the "requirements.txt". 
To run the project locally you have to use git bash or cmd, navigate to your current directory and run the command : "python manage.py runserver:PORT" 
if you don't assign a port it will use the default.